typecasted math.pow to int

+++ _05_FunctionsAndMethods/_10_BinaryToDecimal.java
@@ -1,14 +1,14 @@
 package _05_FunctionsAndMethods;
 
 public class _10_BinaryToDecimal {
-    public static double binaryToDecimal(int binary) {
+    public static int binaryToDecimal(int binary) {
 
-        double decimal = 0;
-        double pow = 0;
+        int decimal = 0;
+        int pow = 0;
 
         while (binary > 0) {
             int ld = binary % 10;
-            decimal = decimal + ld * Math.pow(2, pow);
+            decimal = decimal + ld * (int) Math.pow(2, pow);
             pow++;
             binary /= 10;
         }
